Настраиваю сеть

+++ backpropagation/neuronX.cs
@@ -12,7 +12,7 @@ namespace backpropagation
         {
             x_in = signal;
         }
-        public float out_x() { return x_in ? 1.0f : 0.0f; }
+        public double out_x() { return x_in ? 1.0 : 0.0; }
         private bool x_in;
     }
 }
